backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.1

# NLP (lightweight)
nltk==3.8.1
